show loader when sorting, refactor sorting popup
